Tests/LibWeb/Text/expected/caret-position-ignores-negative-z-decoration.txt

caret below heading moves to following text: true
